Deduplicate InMemoryTextResource
