+++ packages/bulk-edit/src/browser/bulk-edit-frontend-module.ts
@@ -15,13 +15,12 @@
  ********************************************************************************/
 import { ContainerModule } from '@theia/core/shared/inversify';
 import { WidgetFactory } from '@theia/core/lib/browser/widget-manager';
-import { BulkEditTreeWidget, BULK_EDIT_TREE_WIDGET_ID, createBulkEditTreeWidget, InMemoryTextResourceResolver } from './bulk-edit-tree';
+import { BulkEditTreeWidget, BULK_EDIT_TREE_WIDGET_ID, createBulkEditTreeWidget } from './bulk-edit-tree';
 import { FrontendApplicationContribution, LabelProviderContribution } from '@theia/core/lib/browser';
 import { bindViewContribution } from '@theia/core/lib/browser';
 import { BulkEditContribution } from './bulk-edit-contribution';
 import { TabBarToolbarContribution } from '@theia/core/lib/browser/shell/tab-bar-toolbar';
 import { BulkEditTreeLabelProvider } from './bulk-edit-tree-label-provider';
-import { ResourceResolver } from '@theia/core/lib/common';
 import '../../src/browser/style/bulk-edit.css';
 
 export default new ContainerModule(bind => {
@@ -38,7 +37,4 @@ export default new ContainerModule(bind => {
 
     bind(BulkEditTreeLabelProvider).toSelf().inSingletonScope();
     bind(LabelProviderContribution).toService(BulkEditTreeLabelProvider);
-
-    bind(InMemoryTextResourceResolver).toSelf().inSingletonScope();
-    bind(ResourceResolver).toService(InMemoryTextResourceResolver);
 });
