+++ packages/bulk-edit/src/browser/bulk-edit-tree/index.ts
@@ -19,4 +19,3 @@ export * from './bulk-edit-tree-model';
 export * from './bulk-edit-node-selection';
 export * from './bulk-edit-tree-widget';
 export * from './bulk-edit-tree-container';
-export * from './in-memory-text-resource';
