+++ packages/search-in-workspace/src/browser/search-in-workspace-frontend-module.ts
@@ -22,11 +22,9 @@ import { SearchInWorkspaceServer, SIW_WS_PATH } from '../common/search-in-worksp
 import {
     WebSocketConnectionProvider, WidgetFactory, createTreeContainer, TreeWidget, bindViewContribution, FrontendApplicationContribution, LabelProviderContribution
 } from '@theia/core/lib/browser';
-import { ResourceResolver } from '@theia/core';
 import { SearchInWorkspaceWidget } from './search-in-workspace-widget';
 import { SearchInWorkspaceResultTreeWidget } from './search-in-workspace-result-tree-widget';
 import { SearchInWorkspaceFrontendContribution } from './search-in-workspace-frontend-contribution';
-import { InMemoryTextResourceResolver } from './in-memory-text-resource';
 import { SearchInWorkspaceContextKeyService } from './search-in-workspace-context-key-service';
 import { TabBarToolbarContribution } from '@theia/core/lib/browser/shell/tab-bar-toolbar';
 import { bindSearchInWorkspacePreferences } from './search-in-workspace-preferences';
@@ -56,9 +54,6 @@ export default new ContainerModule(bind => {
         const client = ctx.container.get(SearchInWorkspaceClientImpl);
         return WebSocketConnectionProvider.createProxy(ctx.container, SIW_WS_PATH, client);
     }).inSingletonScope();
-
-    bind(InMemoryTextResourceResolver).toSelf().inSingletonScope();
-    bind(ResourceResolver).toService(InMemoryTextResourceResolver);
 
     bindSearchInWorkspacePreferences(bind);
